+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## not released
 
+## v0.3.0 (2024-01-02)
+
+- Fix: #4 Incorrect excerpt on new lines
+- Fix: Sometimes the first image was not used for the preview
+- Add: Additional variables added, `{{url}}`, `{{createdTime}}` and `{{updatedTime}}` #6
+- Improved: Some css changes
+- Improved: Regenerate thumbnail on resource change #5
+
 ## v0.2.0 (2023-12-28)
 
 - Add: Option for a thumbnail